requirements.txt
attrs==21.2.0
autopep8==1.5.7
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
google-api-core==2.0.1
google-api-python-client==2.19.1
google-auth==2.0.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
iniconfig==1.1.1
lark-parser==0.10.1
packaging==21.0
pluggy==1.0.0
protobuf==3.17.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyparsing==2.4.7
pytest==6.2.5
python-hcl2==3.0.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
